# perf(cubesql): Improve rules loading perf