**Check List**
- [x] Tests has been run in packages where changes made if available
- [x] Linter has been run for changed code
- [x] Tests for the changes have been added if not covered yet
- [x] Docs have been added / updated if required

**Description of Changes Made (if issue reference is not provided)**

* Add separate cheaper error type, without backtrace and input part
* Avoid run-time formatting to add `:`
* Avoid calls to replace when `strip_prefix` + `to_string` is enough

On my laptop it gives this:

`cargo bench -- rules_loading`
from ~21.5 ms to ~5.5 ms

`cargo test` reported time for `Running unittests src/lib.rs`
from ~49s to ~35s

`cargo test --release` reported time for `Running unittests src/lib.rs`
from ~6.9s to ~3.5s